+++ qchem/OrcaCalculation.py
@@ -1,10 +1,8 @@
 import os
 import subprocess
-from textwrap import indent
-import comm
-from distutils import core
 from qchem.XYZFile import XYZFile
 from qchem.Molecule import Molecule
+from .OrcaInputFile import OrcaInputFile
 
 class OrcaCalculation:
     """Class capable of running an Orca Calculation"""
@@ -24,6 +22,8 @@ class OrcaCalculation:
     Cores : int
     """Number of Cores to use for the Calculation"""
 
+
+
     OutputFilePath: str
     """The Path to the Output File on the Device"""
 
@@ -33,51 +33,78 @@ class OrcaCalculation:
     OrcaCachePath: str
     """The Path to the Orca Cache on the Local Device"""
 
+    InputFile: OrcaInputFile
+    """The Input File for the Calculation that will be run"""
+
     Index: int
     """Container Index so that they can be run in parallel"""
     
     IsLocal: bool
     """Boolean Flag determining if the Calculation should be run locally or inside a Container"""
 
-    def __init__(self, molecule: Molecule, calculationType: str = None, basisSet: str = None, densityFunctional: str = None, cores : int = 1, index: int = 1):
+    def __init__ (self, inputFile: OrcaInputFile, index: int = 1, isLocal: bool = False):
         
-        if (molecule and isinstance(molecule, (Molecule))):
-            self.CalculationMolecule = molecule
-        else:
-            raise ValueError("Molecule is not defined or is Not of Type Molecule")
+        # Value Type Checking
+        if (not isinstance(inputFile, (OrcaInputFile))):
+            raise ValueError("Input File must be a OrcaInputFile")
         
-        if (isinstance(cores, (int))):
-            self.Cores = cores
-        else:
-            raise ValueError("Cores must be an Integer")
-
-        if (calculationType):
-            if (isinstance(calculationType, (str))):
-                self.CalculationType = calculationType
-            else:
-                raise ValueError("Calculation Type must be a String")
-            
-        if (basisSet):
-            if (isinstance(basisSet, (str))):
-                self.BasisSet = basisSet
-            else:
-                raise ValueError("Basis Set Type must be a String")
-            
-        if (densityFunctional):
-            if (isinstance(densityFunctional, (str))):
-                self.DensityFunctional = densityFunctional
-            else:
-                raise ValueError("Density Functional Type must be a String")
-            
-        if (isinstance(index, (int))):
-            self.Index = index
-        else:
+        if (not isinstance(index, (int))):
             raise ValueError("Index must be an integer")
         
+        if (not isinstance(isLocal, (bool))):
+            raise ValueError("IsLocal must be a boolean")
+        
+        self.InputFile = inputFile
+        self.Index = index
+        self.IsLocal = isLocal
         orcaCache = "OrcaCache"
         self.OrcaCachePath = f'{os.getcwd()}\\{orcaCache}\\{self.CalculationMolecule.Name.replace('.', '')}'
         self.OutputFilePath = f'{self.OrcaCachePath}\\{self.GetOutputFileName()}'
         self.InputFilePath = f'{self.OrcaCachePath}\\{self.GetInputFileName()}'
+
+
+    #def __init__(self, molecule: Molecule, calculationType: str = "", basisSet: str = "", densityFunctional: str = "", cores : int = 1, index: int = 1, isLocal: bool = False):
+    #    
+    #    
+    #    if (not (molecule and isinstance(molecule, (Molecule)))):
+    #        raise ValueError("Molecule is not defined or is Not of Type Molecule")
+    #    
+    #    if (not isinstance(cores, (int))):
+    #        raise ValueError("Cores must be an Integer")
+#
+    #    if (calculationType and not isinstance(calculationType, (str))):
+    #        if (isinstance(calculationType, (str))):
+    #            self.CalculationType = calculationType
+    #        else:
+    #            raise ValueError("Calculation Type must be a String")
+    #        
+    #    if (basisSet):
+    #        if (isinstance(basisSet, (str))):
+    #            self.BasisSet = basisSet
+    #        else:
+    #            raise ValueError("Basis Set Type must be a String")
+    #        
+    #    if (densityFunctional):
+    #        if (isinstance(densityFunctional, (str))):
+    #            self.DensityFunctional = densityFunctional
+    #        else:
+    #            raise ValueError("Density Functional Type must be a String")
+    #        
+    #    if (not isinstance(index, (int))):
+    #        raise ValueError("Index must be an integer")
+    #    
+    #    if (not isinstance(isLocal, (bool))):
+    #        raise ValueError("IsLocal must be a boolean")
+    #    
+    #    
+    #    self.Index = index
+    #    self.Cores = cores
+    #    self.CalculationMolecule = molecule
+    #    self.IsLocal = isLocal
+    #    orcaCache = "OrcaCache"
+    #    self.OrcaCachePath = f'{os.getcwd()}\\{orcaCache}\\{self.CalculationMolecule.Name.replace('.', '')}'
+    #    self.OutputFilePath = f'{self.OrcaCachePath}\\{self.GetOutputFileName()}'
+    #    self.InputFilePath = f'{self.OrcaCachePath}\\{self.GetInputFileName()}'
 
     def RunCalculation(self):
         """Runs a Orca Calculation in a Docker Container """
@@ -91,17 +118,18 @@ class OrcaCalculation:
             os.makedirs(self.OrcaCachePath)
 
         # Save the Input File to the folder
-        self.SaveInputFile(self.OrcaCachePath)
+        self.InputFile.SaveInputFile(self.OrcaCachePath)
+        #self.SaveInputFile(self.OrcaCachePath)
 
         if (self.IsLocal):
-            result = self.RunLocally(orcaCachePath)
+            result = self.RunLocally(self.OrcaCachePath)
         else:
-            result = self.RunDockerContainer(orcaCachePath)
+            result = self.RunDockerContainer(self.OrcaCachePath)
             
         if (result.stderr.__len__() > 0):
             print(f"WARNING Errors Maybe Occured : \n\n{result.stderr}")
 
-        print(f"Calculation Complete : {self.GetInputFileName()}")
+        print(f"Calculation Complete : {self.GetInputFileName()}")  
 
         # Open the Output File and Grab the Content
         with open(self.OutputFilePath, 'r') as file:
@@ -111,10 +139,12 @@ class OrcaCalculation:
         # Create the Command String
         command = ""
         
+        # Windows OS
         if os.name == 'nt':
             orcaPath = subprocess.run("where orca", shell=True, text=True, capture_output=True).stdout.strip(" \n\"").removesuffix(".exe")
             command = f"cd /d {cachePath} && \"{orcaPath}\" \"{self.GetInputFileName()}\" > \"{self.GetOutputFileName()}\""
         else:
+            # Unix based OS (Linux, Mac)
             command = f"cd \"{cachePath}\" && /Orca/orca {self.GetInputFileName()} > {self.GetOutputFileName()}"
         
         # Run the Orca Calculation locally
@@ -139,90 +169,60 @@ class OrcaCalculation:
         
         return result
 
+
     def GetInputFileName (self):
         """Returns the Input File Name with it's extension"""
-        return f"{self.CalculationMolecule.Name}.inp"
+        return f"{self.InputFile.Name}.inp"
 
     def GetOutputFileName (self):
         """Returns the Output File Name with it's extension"""
-        return f"{self.CalculationMolecule.Name}.out"
+        return f"{self.InputFile.Name}.out"
 
-    def GetInputFile (self):
-        """Generates a Input file in String format"""
+    #def GetInputFileName (self):
+    #    """Returns the Input File Name with it's extension"""
+    #    return f"{self.CalculationMolecule.Name}.inp"
+#
+    #def GetOutputFileName (self):
+    #    """Returns the Output File Name with it's extension"""
+    #    return f"{self.CalculationMolecule.Name}.out"
 
-        firstLine = "!"
+    #def GetInputFile (self):
+    #    """Generates a Input file in String format"""
+#
+    #    firstLine = "!"
+#
+    #    # Check if the Properties are defined and Add them to the First line of the Input File
+    #    if (self.DensityFunctional):
+    #        firstLine += f"{self.DensityFunctional}"
+#
+    #    if (self.BasisSet):
+    #        firstLine += f" {self.BasisSet}"
+#
+    #    if (self.CalculationType):
+    #        firstLine += f" {self.CalculationType}"
+#
+    #    if (self.Cores > 1):
+    #        if (self.Cores < 9):
+    #            firstLine += " " + f"PAL{self.Cores}"
+    #        else:
+    #            firstLine += f"\n%PAL NPROCS {self.Cores} END"
+#
+    #    # Generate XYZ Wrappers
+    #    xyzWrapperStart = "* xyz 0 1"
+    #    xyzWrapperEnd = "*"
+#
+    #    # Create a XYZ File for the Molecule
+    #    xyz = XYZFile(self.CalculationMolecule)
+#
+    #    # Generate the Entire Input File as a String
+    #    inputFile = f"{firstLine}\n{xyzWrapperStart}\n{xyz.GetXYZBody()}\n{xyzWrapperEnd}"
+#
+    #    return inputFile
 
-        # Check if the Properties are defined and Add them to the First line of the Input File
-        if (self.DensityFunctional):
-            firstLine += f"{self.DensityFunctional}"
-
-        if (self.BasisSet):
-            firstLine += f" {self.BasisSet}"
-
-        if (self.CalculationType):
-            firstLine += f" {self.CalculationType}"
-
-        if (self.Cores > 1):
-            if (self.Cores < 9):
-                firstLine += " " + f"PAL{self.Cores}"
-            else:
-                firstLine += f"\n%PAL NPROCS {self.Cores} END"
-
-        # Generate XYZ Wrappers
-        xyzWrapperStart = "* xyz 0 1"
-        xyzWrapperEnd = "*"
-
-        # Create a XYZ File for the Molecule
-        xyz = XYZFile(self.CalculationMolecule)
-
-        # Generate the Entire Input File as a String
-        inputFile = f"{firstLine}\n{xyzWrapperStart}\n{xyz.GetXYZBody()}\n{xyzWrapperEnd}"
-
-        return inputFile
-
-    def SaveInputFile (self, filePath: str):
-        """Saves a Input File using the Settings Provided to the Path Specified"""
-        with open(os.path.join(filePath, self.GetInputFileName()), "w") as file:
-            file.write(self.GetInputFile())
-
-class OrcaInputFile:
-    # Template is an input file with missing variables in the form of &{variable_name}
-    # kwargs is a dictionary with variable names as keys and their values as values
-    # Example:
-    # !SP &{basis} PBE
-    # *xyzfile 0 1 aspirin.xyz
-    # with variables={'basis': 'def2-SVP'}
-    # will be converted to 
-    # !SP def2-SVP PBE
-    # *xyzfile 0 1 aspirin.xyz
-
-    # Example with template file:
-    # tester = OrcaInputFile(OrcaInputTemplate.BASIC
-    #   , calculation='OPT'
-    #   , basis='def2-SVP'
-    #   , functional='PBE'
-    #   , xyzfile='aspirin.xyz'
-    #)
-    
-    def __init__(self, template: str, **variables):
-        self.template = template
-        self.variables = variables
-        self.inputfile = self.GenerateInputFile()
-
-    def GenerateInputFile(self) -> str:
-        """Generates the input file content by replacing placeholders with actual values."""
-        if isinstance(self.template, OrcaInputTemplate):
-            input_content=self.template.value    
-        else:
-            with open(self.template, 'r') as file:
-                input_content = file.read()
-        
-        for key, value in self.variables.items():
-            placeholder = f'&{{{key}}}'
-            input_content = input_content.replace(placeholder, str(value))
-        return input_content
-
-    def SaveInputFile(self, file_path: str):
-        """Saves the generated input file content to a specified path."""
-        with open(file_path, 'w') as file:
-            file.write(self.inputfile)
+    #def SaveInputFile (self, filePath: str):
+    #    """Saves a Input File using the Settings Provided to the Path Specified"""
+    #    
+    #    
+    #    
+    #    with open(os.path.join(filePath, self.GetInputFileName()), "w") as file:
+    #        file.write(self.InputFile.InputFile)
